Further tidy up of the img tile interface.
