+++ lib/cartopy/examples/gmaptiles.py
@@ -15,105 +15,74 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with cartopy.  If not, see <http://www.gnu.org/licenses/>.
 
-# XXX Needs a polish. The new image tile support should simplify.
+
+# PLEASE NOTE: This example is a work in progress. It is anticipated that a tile generating 
+# interface will simplify this significantly.
 
 import matplotlib.pyplot as plt
 import numpy
 
 import cartopy
 from cartopy.examples.waves import sample_data
+import cartopy.io.img_tiles
 
-# XXX Turn into an object
-def tile_bbox(prj, x, y, z, lat_extent_at_z0=(-85., 85.), bottom_up=True, native=True):
-    """
-    Returns the x0, x1, y0, y1 bounding box for the given x, y, z tile position. 
-    """
-    n = 2**z
-    assert 0 <= x <= (n - 1), "Tile's x index is out of range. Upper limit %s. Got %s" % (n, x)
-    assert 0 <= y <= (n - 1), "Tile's y index is out of range. Upper limit %s. Got %s" % (n, y)
-    
-    x0 = -180.
-    # compute the box height in native coordinates for this zoom level
-    box_w = 360. / n
-    
-    result = prj.transform_points(cartopy.crs.PlateCarree(), numpy.array([0., 0]), numpy.array(lat_extent_at_z0))
-    y1, y0 = result[:, 1]
-    # compute the box height in native coordinates for this zoom level
-    box_h = (y1 - y0) / n
-    
-    # compute the native x & native y extents of the box
-#    x_lower, x_upper = x0 + x*box_w, x0 + (x+1)*box_w
-#    y_lower, y_upper = y0 + y*box_h, y0 + (y+1)*box_h
-    n_xs = x0 + (x + numpy.arange(0, 2, dtype=numpy.float64)) * box_w
-    n_ys = y0 + (y + numpy.arange(0, 2, dtype=numpy.float64)) * box_h
-    
-    if not bottom_up:
-        # n.b. assumes that the projection is symmetric
-        n_ys = -1 * n_ys[::-1]
-        
-    if native:
-        return n_xs, n_ys
-    else:
-        b_xs, b_ys = prj.unproject_points(n_xs, n_ys)
-    
-        return b_xs, b_ys
 
 def produce_axes2(ax):
     coords = [(-0.08, 51.53), (132.00, 43.17)] # London to Vladivostock
     r = ax.plot(*zip(*coords), transform=ll)
-    
+
     x, y, z = sample_data()
     ax.contour(x, y, z, transform=ll)
     ax.coastlines()
     ax.gridlines()
-    
+
 def produce_axes(ax):
     import cartopy
     import cartopy.mpl_integration.patch as pt
     import matplotlib.pyplot as plt
     import matplotlib.path as mpath
     import numpy
-    
+
     import matplotlib.textpath
     import matplotlib.patches
     from matplotlib.font_manager import FontProperties
-    
+
     pc = cartopy.prj.PlateCarree()
-    
+
     # create a transform from PlateCarree to the axes' projection
     pc_t = pc._as_mpl_transform(ax)
-    
+
     ax.coastlines()
-    ax.gridlines()
+    ax.gridlines(15)
     im = ax.bluemarble()
-    
+
     name, pos = 'cartopy', (-180, -30)
     logo_path = matplotlib.textpath.TextPath(pos, name, size=1, prop=FontProperties(family='Arial', weight='bold'))
     # put the letters in the right place
     logo_path._vertices *= numpy.array([95, 160])
-    
-    im.set_clip_path(logo_path, transform=pc_t)    
-    
+
+    im.set_clip_path(logo_path, transform=pc_t)
+
 
 if __name__ == '__main__':
     prj = cartopy.prj.Mercator()
     ll = cartopy.prj.PlateCarree()
-    
+
     # set up a 1:1 ratio figure, which when saved will produce a 256x256 image
     plt.figure(figsize=[4, 4])
     ax = plt.axes([0, 0, 1, 1], projection=prj, frameon=False)
-    
+
     ax.get_xaxis().set_visible(False)
     ax.get_yaxis().set_visible(False)
-    
+
     produce_axes(ax)
-    
+
     ax.patch.set_visible(False)
     ax.outline_patch.set_visible(False)
     for z in range(0, 4):
-        for i in numpy.arange(0, 2**z):
-            for j in xrange(0, 2**z):
-                x_lim, y_lim = tile_bbox(prj, i, j, z, bottom_up=False)
+        for i in numpy.arange(0, 2 ** z):
+            for j in xrange(0, 2 ** z):
+                x_lim, y_lim = cartopy.io.img_tiles.GoogleTiles.tile_bbox(prj, i, j, z, bottom_up=False)
                 print i, j, z, y_lim
                 ax.set_xlim(x_lim)
                 ax.set_ylim(y_lim)
@@ -122,5 +91,5 @@ if __name__ == '__main__':
                 # (http://old.nabble.com/Transparency-with-fig.canvas.mpl_connect-td27724532.html)
                 plt.gcf().canvas.get_renderer().clear()
                 plt.savefig('logo/z%s_y%s_x%s.png' % (z, j, i), dpi=64, transparent=True)
-              
+
 #    plt.show()
